requirements.txt
pyshark>=0.6
jinja2>=3.1.2
requests>=2.31.0
python-dotenv>=1.0.1
